Refactor using Externals and Linq, Callout extended to use BaseScript
Rewrote some of the stuff to use Externals, since it's shorter and easier to read. Also rewrote a couple of things using Linq, because who doesn't love turning 15 lines into 3. Removed some unused Usings. Moved around some of the stuff in WorldZone; the way I've done it seems a lot easier, but idk if you guys wrote it like that for a specific reason.